Move the BulkProcessor feature to the framework
