+++ framework/src/main/java/fr/pilato/elasticsearch/crawler/fs/framework/bulk/FsCrawlerAdvancedBulkProcessorListener.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.crawler.fs.thirdparty.wpsearch;
+package fr.pilato.elasticsearch.crawler.fs.framework.bulk;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
