+++ framework/src/main/java/fr/pilato/elasticsearch/crawler/fs/framework/bulk/FsCrawlerOperation.java
@@ -17,24 +17,7 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.crawler.fs.thirdparty.wpsearch;
+package fr.pilato.elasticsearch.crawler.fs.framework.bulk;
 
-import java.util.ArrayList;
-import java.util.List;
-
-abstract public class FsCrawlerBulkRequest<T extends FsCrawlerOperation<T>> {
-
-    private final List<T> operations = new ArrayList<>();
-
-    public int numberOfActions() {
-        return operations.size();
-    }
-
-    public void add(T request) {
-        operations.add(request);
-    }
-
-    public List<T> getOperations() {
-        return operations;
-    }
+public interface FsCrawlerOperation<T extends FsCrawlerOperation<T>> extends Comparable<T> {
 }
